Fixed indentation and get() error handler
I fixed the indentation and the one-line conditional loops, some of them were all on 1 line and some where on 2 lines, I put them all on 2 lines for readability and consistency's sake. I also implemented the error handler for the get() function on the issue that's linked to this one.